Add Classifier base class
This commit adds a classifier base class that
can be sub-classed in order to create new classifiers.
It contains all the basic methods like setting the train
and test dataset, training an algorithm, evaluating an
algorithm, loading and saving models from/to disk and
of course predicting/classifying a new instance.