Dataset statistics enhancements and fixes
This commit fixes a bug in Statistics where all attributes
were recognized as NOMINAL. Now attributes are correctly recognized
as either NOMINAL or NUMERIC in order to calculate different statistics
for each type of attribute.
This commit also adds a test case which loads a csv file from disk
containing both NUMERIC and NOMINAL attributes and checks to see
if Statistics parse the Dataset correctly.